requirements_appnotion.txt
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
matplotlib==3.7.1
numpy==1.26.4
diffusers==0.32.2
transformers==4.49.0